# Improve memory performance of Queue implementation