@mbleigh 2018-11-12
@bkendall the deletion works by fanning out recursively and doing shallow reads to determine Node size, so it's inherently a queue that grows asynchronously. It's not surprising to me that this would change the memory footprint significantly for that case.

Hosting doesn't really matter since all files are listed up front.