@fredzqm 2018-11-12
@bkendall The issue was not caused by the velocity in which tasks are processed, but by the total number of tasks. If the concurrency limit is reached, `chunkDelete` will stop sending requests until one active request finishes.

If the tree structure is relatively flat, then it will end up scanning the subtress in chunks, and delete them. In this specific case, releasing completed tasks will limit tasks queue size to batchSize*concurrency.

This is a quite common use case, so I figure I should fix it.