@bkendall 2018-11-12
Got it. That's great. Thanks for clearing it up!

On Mon, Nov 12, 2018 at 11:01 AM Qinmao Zhang <notifications@github.com>
wrote:

> @bkendall <https://github.com/bkendall> The issue was not caused by the
> velocity in which tasks are processed, but by the total number of tasks. If
> the concurrency limit is reached, chunkDelete will stop sending requests
> until one active request finishes.
>
> If the tree structure is relatively flat, then it will end up scanning the
> subtress in chunks, and delete them. In this specific case, releasing
> completed tasks will limit tasks queue size to batchSize*concurrency.
>
> This is a quite common use case, so I figure I should fix it.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/firebase/firebase-tools/pull/1011#issuecomment-437974336>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAJx7B-QXyZPtvsfqsccaJHi0cSU1C5sks5uubcagaJpZM4YY93c>
> .
>


-- 
Bryan Kendall | Software Engineer | bkend@google.com